refactor: remove component barrels
