--- components/article-dates/index.ts
@@ -1,1 +1,0 @@
-export { ArticleDates } from './article-dates';
