--- components/article-meta/index.ts
@@ -1,1 +1,0 @@
-export { ArticleMeta } from './article-meta';
