--- components/blog-post-preview/index.ts
@@ -1,1 +1,0 @@
-export * from './blog-post-preview';
