--- components/blog-post-square-preview/index.ts
@@ -1,1 +1,0 @@
-export * from './blog-post-square-preview';
