--- components/bookmark-section/index.ts
@@ -1,1 +1,0 @@
-export { BookmarkSection } from './bookmark-section';
