+++ components/categories.test.tsx
