+++ components/categories.tsx
