--- components/floating-reactions/index.ts
@@ -1,1 +1,0 @@
-export { FloatingReactions } from './floating-reactions';
