--- components/footer-link/index.ts
@@ -1,1 +1,0 @@
-export * from './footer-link';
