--- components/header-link/index.ts
@@ -1,1 +1,0 @@
-export * from './header-link';
