--- components/linkedin-feedback/index.ts
@@ -1,1 +1,0 @@
-export { LinkedinFeedback } from './linkedin-feedback';
