+++ components/mdx-components/mdx-components.tsx
@@ -1,6 +1,6 @@
 import Image, { ImageProps } from 'next/image';
 import Link from 'next/link';
-import React, { AnchorHTMLAttributes, ClassAttributes, JSX } from 'react';
+import { AnchorHTMLAttributes, ClassAttributes, JSX } from 'react';
 
 import { Routes } from '@/lib/routes';
