--- components/mobile-menu/index.ts
@@ -1,1 +1,0 @@
-export * from './mobile-menu';
