--- components/no-results/index.ts
@@ -1,1 +1,0 @@
-export { NoResults } from './no-results';
