+++ components/resource-card.tsx
