--- components/resource-card/index.ts
@@ -1,1 +1,0 @@
-export { ResourceCard, type ResourceCardProps } from './resource-card';
