--- components/search-input/index.ts
@@ -1,1 +1,0 @@
-export { SearchInput } from './search-input';
