--- components/share-button/index.ts
@@ -1,4 +1,0 @@
-export * from './facebook-share';
-export * from './linkedin-share';
-export * from './twitter-share';
-export * from './telegram-share';
