--- components/table-of-content/index.ts
@@ -1,1 +1,0 @@
-export { TableOfContent } from './table-of-content';
