+++ components/view-counter.test.tsx
