+++ components/view-counter.tsx
