--- components/view-counter/index.ts
@@ -1,1 +1,0 @@
-export * from './view-counter';
