--- components/year-separator/index.ts
@@ -1,1 +1,0 @@
-export { YearSeparator } from './year-separator';
